Forgot to returning owner in query

+++ src/services/postgres/NotesServices.js
@@ -71,7 +71,7 @@ class NoteServices {
     async editNoteById(id, {title, body, tags}){
         const updatedAt = new Date().toISOString();
         const query = {
-            text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id',
+            text: 'UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner',
             values: [title, body, tags, updatedAt, id],
         };
 
@@ -88,7 +88,7 @@ class NoteServices {
 
     async deleteNoteById(id) {
         const query = {
-            text: 'DELETE FROM notes WHERE id = $1 RETURNING id',
+            text: 'DELETE FROM notes WHERE id = $1 RETURNING id, owner',
             values: [id],
         }
